FEAT: Cadastro do cliente realizado com sucesso, ao inserir os dados corretamente no postman o usuário já fica salvo no banco de dados
